feat(health): promote Triage view into @repowise-dev/ui
Move the composed Code Health Triage surface (KPI strip, master-detail
map + inspector rail, fix-next queue, findings sidebar, hot-function and
hidden-coupling panels) out of packages/web into a shared TriageView in
packages/ui/src/health, behind a CodeHealthAdapter that injects the data
fetchers, link builders, navigation, and file-detail drawer as props.

Web's triage-tab.tsx is now a thin host that binds the adapter to its
/api client, /repos/:id routing, and HealthFileDrawerHost; the rendered
output is unchanged. Adds swr as a ui peerDependency so the shared view
can own its fetch orchestration off the consuming app's swr instance.
