+++ packages/ui/package.json
@@ -149,7 +149,8 @@
   "peerDependencies": {
     "next-themes": "^0.4.0",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "swr": "^2.2.5"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.5.0",
