+++ packages/ui/src/health/index.ts
@@ -36,3 +36,5 @@ export * from "./hot-functions-panel";
 export * from "./hidden-coupling-list";
 export * from "./recalibration-banner";
 export * from "./defect-accuracy-card";
+export * from "./code-health-adapter";
+export * from "./triage-view";
